requirements.txt
pandas
pandas-ta
pyyaml
websockets
requests
pytest
